feat: added roles to application
