+++ src/main/java/com/example/survey/controller/RegisterRequest.java
@@ -7,6 +7,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.validation.constraints.NotBlank;
+import javax.validation.constraints.Size;
 
 @Data
 @Builder
@@ -15,13 +16,18 @@ import javax.validation.constraints.NotBlank;
 @UniqueUser
 public class RegisterRequest {
 
+	@Size(min = 5, max = 50)
 	private String firstName;
+
+	@Size(min = 5, max = 50)
 	private String lastName;
 
 	@NotBlank
+	@Size(min = 5, max = 50)
 	private String password;
 
 	@NotBlank
+	@Size(min = 5, max = 50)
 	private String username;
 
 }
