+++ src/main/java/com/example/survey/entities/UserRoles.java
@@ -6,6 +6,8 @@ import org.springframework.security.core.GrantedAuthority;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
@@ -23,6 +25,7 @@ public class UserRoles implements GrantedAuthority {
 	private Long id;
 
 	@Column(name = "usp_role")
+	@Enumerated(EnumType.STRING)
 	private Roles role;
 
 	@Override
